web/university-of-straya-part1/src/ass/requirements.txt
flask
requests
pyjwt[crypto]
Flask-Caching
Flask-SQLAlchemy
python-magic
bcrypt